UCA 6.0 data, test data, and bug fixes; from branches/markus/uca60 -r 28826:28857
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/7264


git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@28875 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/i18n/ucol_res.cpp
@@ -397,9 +397,9 @@ ucol_openRules( const UChar        *rules,
         /* so something might be done here... or on lower level */
 #ifdef UCOL_DEBUG
         if(*status == U_ILLEGAL_ARGUMENT_ERROR) {
-            fprintf(stderr, "bad option starting at offset %i\n", src.current-src.source);
+            fprintf(stderr, "bad option starting at offset %i\n", (int)(src.current-src.source));
         } else {
-            fprintf(stderr, "invalid rule just before offset %i\n", src.current-src.source);
+            fprintf(stderr, "invalid rule just before offset %i\n", (int)(src.current-src.source));
         }
 #endif
         goto cleanup;
